requirements.txt
```pip-requirements
beautifulsoup4>=4.12.2
lxml>=4.9.3
requests>=2.31.0
httpx>=0.27.0
tenacity>=8.2.3
tqdm>=4.66.4
fastapi>=0.111.0
uvicorn>=0.30.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.1.0
transformers>=4.35.2
streamlit>=1.33.0
```